requirements.txt
anthropic>=0.52.0
python-telegram-bot>=21.0
groq>=0.11.0
